refactor(ResultView): move to result view
