+++ src/main/kotlin/lotto/Main.kt
@@ -1,20 +1,17 @@
 import lotto.domain.LottoItems
 import lotto.view.InputView
+import lotto.view.ResultView
 
 fun main() {
     val inputMoney = InputView.inputMoney()
-    val buyCount = inputMoney / 1000
-    println("${buyCount}개를 구매했습니다.")
+    val buyCount = InputView.getBuyCount(inputMoney)
+
+    ResultView.printBuyCount(buyCount)
 
     val lottoItems = LottoItems(buyCount)
     val lottos = lottoItems.getLottoItems()
 
     val winningNumbers = InputView.inputWinningNumber()
 
-    println(lottos.joinToString(""))
-
-    println("당첨 통계")
-    println("---------")
-    val result = lottoItems.getWinLottos(winningNumbers)
-    println(result.joinToString(""))
+    ResultView.printResult(lottos, lottoItems, winningNumbers)
 }
