+++ src/main/kotlin/lotto/view/InputView.kt
@@ -3,10 +3,11 @@ package lotto.view
 import lotto.domain.value.LottoNumber
 
 object InputView {
-    const val INPUT_MONEY_MESSAGE = "구입금액을 입력해 주세요."
-    const val INPUT_NOT_INTEGER_ERROR_MESSAGE = "숫자만 넣어 주세요"
-    const val INPUT_WINNING_NUMBER_MESSAGE = "지난 주 당첨 번호를 입력해 주세요."
-    const val INPUT_NOT_VALID_ERROR_MESSAGE = "콤마(,)로 연결된 6자리 숫자를 입력하세요."
+    private const val INPUT_MONEY_MESSAGE = "구입금액을 입력해 주세요."
+    private const val INPUT_NOT_INTEGER_ERROR_MESSAGE = "숫자만 넣어 주세요"
+    private const val INPUT_WINNING_NUMBER_MESSAGE = "지난 주 당첨 번호를 입력해 주세요."
+    private const val INPUT_NOT_VALID_ERROR_MESSAGE = "콤마(,)로 연결된 6자리 숫자를 입력하세요."
+    private const val PRICE_PER_UNIT = 1000
 
     fun inputMoney(): Int {
         println(INPUT_MONEY_MESSAGE)
@@ -19,4 +20,6 @@ object InputView {
             LottoNumber(it.toInt())
         } ?: throw IllegalArgumentException(INPUT_NOT_VALID_ERROR_MESSAGE)
     }
+
+    fun getBuyCount(inputMoney: Int) = inputMoney / PRICE_PER_UNIT
 }
